frontend: refactor Remote MCP create page
